Cancel the RDD persisting operation
I tried to benchmark the conditions with RDD persisting or without RDD
persisting.
Surprisingly, it took longer time if I persisted the RDD. This may be
due to that it takes some time to persist the RDD itself. So, it's only
worthwhile to persist RDD when it's big or it's re-used for very many
time.
Here, I would cancel this operation.