Fix an error
I found a weird error. For some high-ranking packages, in their HTML
files, their ranking is “Inf”. Not sure why. Here I change the
algorithm with which I calculate the ranking. Hope it would work. But
the new algorithm can’t deal with multiple packages with the same count
of downloads. Will try to fix later